+++ test/fixtures/cookbooks/test/recipes/default.rb
@@ -117,7 +117,7 @@ end
     port 80
     weight 5
     inhibit_on_failure true
-    healthcheck resources(keepalived_http_get: 'port-80').path
+    healthcheck resources(keepalived_http_get: 'port-80').config_file
   end
 
   keepalived_real_server "#{addr}-443" do
@@ -125,12 +125,12 @@ end
     port 443
     weight 5
     inhibit_on_failure true
-    healthcheck resources(keepalived_ssl_get: 'port-443').path
+    healthcheck resources(keepalived_ssl_get: 'port-443').config_file
   end
 end
 
 https_servers = %w( 192.168.1.13 192.168.1.14 ).map do |addr|
-  resources(keepalived_real_server: "#{addr}-443").path
+  resources(keepalived_real_server: "#{addr}-443").config_file
 end
 
 keepalived_virtual_server '192.168.1.5 443' do
@@ -142,7 +142,7 @@ keepalived_virtual_server '192.168.1.5 443' do
 end
 
 http_servers = %w( 192.168.1.13 192.168.1.14 ).map do |addr|
-  resources(keepalived_real_server: "#{addr}-80").path
+  resources(keepalived_real_server: "#{addr}-80").config_file
 end
 
 keepalived_virtual_server '192.168.1.5 80' do